fix: emit fake co2 measure every second
